separated main from the business results page (named business_search.dart) and added commits to login -Ivan
